requirements.txt
flask
werkzeug
flask-sqlalchemy
flask-testing
python-slugify
factory_boy
arrow